fix: fix typo and create new file in utils
