+++ src/Utils/Role/index.jsx
@@ -1,0 +1,7 @@
+const Role = () => {
+  const roleString = localStorage.getItem("role");
+  const role = JSON.parse(roleString);
+  return role;
+};
+
+export default Role;
