+++ src/Utils/Token/index.jsx
@@ -1,0 +1,7 @@
+const Token = () => {
+  const tokenString = sessionStorage.getItem("token");
+  const userToken = JSON.parse(tokenString);
+  return userToken;
+};
+
+export default Token;
